report template and guide added
The report template is added with small additions. The report guide is also there so that we can check that we are doing the report correctly. The report will be done online between isaac and Alif, after we have finalised a section only then will the report template be added to